backend/requirements.txt
# FastAPI & Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# ML/DL
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.2

# LLM - API-based providers (recommended)
openai==1.51.0
google-generativeai==0.8.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
